refactor: python standard refactors moved from the reserved lib folder to src update a more robust .gitignore
